+++ test/utils/ApiClient.spec.js
@@ -133,23 +133,29 @@ describe("utils/ApiClient", () => {
                     test: "value",
                 })
                 .reply(404, {
-                    errors: [
-                        "test1",
-                        "test2",
-                    ],
+                    errors: [{
+                        message: "message",
+                        code: "FORBIDDEN",
+                        relatedResources: ["trm-f3d38df1-adb7-4127-9858-e72ebe682a79",
+                            "trm-601b1401-4464-4f3f-97b3-09079ee7723b"],
+                    }],
                 });
 
             client.doPost("test", { test: "value" }, {}, (err, body, res) => {
-                err.should.be.deep.equal([
-                    "test1",
-                    "test2",
-                ]);
+                err.should.be.deep.equal([{
+                    message: "message",
+                    code: "FORBIDDEN",
+                    relatedResources: ["trm-f3d38df1-adb7-4127-9858-e72ebe682a79",
+                        "trm-601b1401-4464-4f3f-97b3-09079ee7723b"],
+                }]);
 
                 body.should.be.deep.equal({
-                    errors: [
-                        "test1",
-                        "test2",
-                    ],
+                    errors: [{
+                        message: "message",
+                        code: "FORBIDDEN",
+                        relatedResources: ["trm-f3d38df1-adb7-4127-9858-e72ebe682a79",
+                            "trm-601b1401-4464-4f3f-97b3-09079ee7723b"],
+                    }],
                 });
 
                 res.status.should.be.equal(404);
@@ -176,10 +182,12 @@ describe("utils/ApiClient", () => {
                     .filteringPath(() => "/")
                     .matchHeader("Authorization", authHeader)
                     .matchHeader("User-Agent", `Hyperwallet Node SDK v${packageJson.version}`)
-                    .matchHeader("Accept", "application/json")
+                    .matchHeader("Accept", "application/jose+json")
                     .matchHeader("Content-Type", "application/jose+json")
                     .post("/", /.+/)
-                    .reply(201, encryptedBody);
+                    .reply(200, encryption.base64Decode(encryptedBody), {
+                        "Content-Type": "application/jose+json",
+                    });
 
                 clientWithEncryption.doPost("test", { message: "Test message" }, {}, (err, body, res) => {
                     expect(err).to.be.undefined();
@@ -333,23 +341,29 @@ describe("utils/ApiClient", () => {
                     test: "value",
                 })
                 .reply(404, {
-                    errors: [
-                        "test1",
-                        "test2",
-                    ],
+                    errors: [{
+                        message: "message",
+                        code: "FORBIDDEN",
+                        relatedResources: ["trm-f3d38df1-adb7-4127-9858-e72ebe682a79",
+                            "trm-601b1401-4464-4f3f-97b3-09079ee7723b"],
+                    }],
                 });
 
             client.doPut("test", { test: "value" }, {}, (err, body, res) => {
-                err.should.be.deep.equal([
-                    "test1",
-                    "test2",
-                ]);
+                err.should.be.deep.equal([{
+                    message: "message",
+                    code: "FORBIDDEN",
+                    relatedResources: ["trm-f3d38df1-adb7-4127-9858-e72ebe682a79",
+                        "trm-601b1401-4464-4f3f-97b3-09079ee7723b"],
+                }]);
 
                 body.should.be.deep.equal({
-                    errors: [
-                        "test1",
-                        "test2",
-                    ],
+                    errors: [{
+                        message: "message",
+                        code: "FORBIDDEN",
+                        relatedResources: ["trm-f3d38df1-adb7-4127-9858-e72ebe682a79",
+                            "trm-601b1401-4464-4f3f-97b3-09079ee7723b"],
+                    }],
                 });
 
                 res.status.should.be.equal(404);
@@ -376,10 +390,10 @@ describe("utils/ApiClient", () => {
                     .filteringPath(() => "/")
                     .matchHeader("Authorization", authHeader)
                     .matchHeader("User-Agent", `Hyperwallet Node SDK v${packageJson.version}`)
-                    .matchHeader("Accept", "application/json")
+                    .matchHeader("Accept", "application/jose+json")
                     .matchHeader("Content-Type", "application/jose+json")
                     .put("/", /.+/)
-                    .reply(201, encryptedBody);
+                    .reply(201, encryption.base64Decode(encryptedBody), { "Content-Type": "application/jose+json" });
 
                 clientWithEncryption.doPut("test", { message: "Test message" }, {}, (err, body, res) => {
                     expect(err).to.be.undefined();
@@ -448,10 +462,10 @@ describe("utils/ApiClient", () => {
                     .filteringPath(() => "/")
                     .matchHeader("Authorization", authHeader)
                     .matchHeader("User-Agent", `Hyperwallet Node SDK v${packageJson.version}`)
-                    .matchHeader("Accept", "application/json")
+                    .matchHeader("Accept", "application/jose+json")
                     .matchHeader("Content-Type", "application/jose+json")
                     .put("/", /.+/)
-                    .reply(201, encryptedBody);
+                    .reply(201, encryption.base64Decode(encryptedBody), { "Content-Type": "application/jose+json" });
 
                 clientWithEncryption.doPut("test", { message: "Test message" }, {}, (err, body, res) => {
                     err.should.be.deep.equal("Failed to decrypt response for PUT request");
@@ -478,7 +492,7 @@ describe("utils/ApiClient", () => {
                 .filteringPath(() => "/")
                 .matchHeader("Authorization", authHeader)
                 .matchHeader("User-Agent", `Hyperwallet Node SDK v${packageJson.version}`)
-                .matchHeader("Accept", "application/json")
+                .matchHeader("Accept", "application/jose+json")
                 .matchHeader("Content-Type", "application/jose+json")
                 .put("/", /.+/)
                 .reply(404, {
@@ -486,7 +500,7 @@ describe("utils/ApiClient", () => {
                         "test1",
                         "test2",
                     ],
-                });
+                }, { "Content-Type": "application/jose+json" });
 
             clientWithEncryption.doPut("test", { message: "Test message" }, {}, (err, body, res) => {
                 err.should.be.deep.equal([
@@ -601,23 +615,29 @@ describe("utils/ApiClient", () => {
                 .matchHeader("Accept", "application/json")
                 .get("/rest/v3/test")
                 .reply(404, {
-                    errors: [
-                        "test1",
-                        "test2",
-                    ],
+                    errors: [{
+                        message: "message",
+                        code: "FORBIDDEN",
+                        relatedResources: ["trm-f3d38df1-adb7-4127-9858-e72ebe682a79",
+                            "trm-601b1401-4464-4f3f-97b3-09079ee7723b"],
+                    }],
                 });
 
             client.doGet("test", {}, (err, body, res) => {
-                err.should.be.deep.equal([
-                    "test1",
-                    "test2",
-                ]);
+                err.should.be.deep.equal([{
+                    message: "message",
+                    code: "FORBIDDEN",
+                    relatedResources: ["trm-f3d38df1-adb7-4127-9858-e72ebe682a79",
+                        "trm-601b1401-4464-4f3f-97b3-09079ee7723b"],
+                }]);
 
                 body.should.be.deep.equal({
-                    errors: [
-                        "test1",
-                        "test2",
-                    ],
+                    errors: [{
+                        message: "message",
+                        code: "FORBIDDEN",
+                        relatedResources: ["trm-f3d38df1-adb7-4127-9858-e72ebe682a79",
+                            "trm-601b1401-4464-4f3f-97b3-09079ee7723b"],
+                    }],
                 });
 
                 res.status.should.be.equal(404);
@@ -644,9 +664,9 @@ describe("utils/ApiClient", () => {
                     .filteringPath(() => "/")
                     .matchHeader("Authorization", authHeader)
                     .matchHeader("User-Agent", `Hyperwallet Node SDK v${packageJson.version}`)
-                    .matchHeader("Accept", "application/json")
+                    .matchHeader("Accept", "application/jose+json")
                     .get("/")
-                    .reply(200, encryptedBody);
+                    .reply(200, encryption.base64Decode(encryptedBody), { "Content-Type": "application/jose+json" });
 
                 clientWithEncryption.doGet("test", {}, (err, body, res) => {
                     expect(err).to.be.undefined();
@@ -687,9 +707,10 @@ describe("utils/ApiClient", () => {
             const rawRes = {
                 body: "test",
                 status: 200,
+                type: "application/json",
             };
 
-            const callback = client.wrapCallback((err, body, res) => {
+            const callback = client.wrapCallback("POST", (err, body, res) => {
                 expect(err).to.be.undefined();
 
                 body.should.be.equal("test");
@@ -705,21 +726,31 @@ describe("utils/ApiClient", () => {
 
             const rawRes = {
                 body: {
-                    errors: [
-                        "test1",
-                        "test2",
-                    ],
+                    errors: [{
+                        message: "message",
+                        code: "FORBIDDEN",
+                        relatedResources: ["trm-f3d38df1-adb7-4127-9858-e72ebe682a79",
+                            "trm-601b1401-4464-4f3f-97b3-09079ee7723b"],
+                    }],
                 },
                 status: 404,
+                type: "application/json",
             };
 
-            const callback = client.wrapCallback((err, body, res) => {
-                err.should.be.deep.equal(["test1", "test2"]);
+            const callback = client.wrapCallback("POST", (err, body, res) => {
+                err.should.be.deep.equal([{
+                    message: "message",
+                    code: "FORBIDDEN",
+                    relatedResources: ["trm-f3d38df1-adb7-4127-9858-e72ebe682a79",
+                        "trm-601b1401-4464-4f3f-97b3-09079ee7723b"],
+                }]);
                 body.should.be.deep.equal({
-                    errors: [
-                        "test1",
-                        "test2",
-                    ],
+                    errors: [{
+                        message: "message",
+                        code: "FORBIDDEN",
+                        relatedResources: ["trm-f3d38df1-adb7-4127-9858-e72ebe682a79",
+                            "trm-601b1401-4464-4f3f-97b3-09079ee7723b"],
+                    }],
                 });
                 rawRes.should.be.deep.equal(res);
 
@@ -734,12 +765,64 @@ describe("utils/ApiClient", () => {
             const rawRes = {
                 body: "test",
                 status: 404,
+                type: "application/json",
             };
 
-            const callback = client.wrapCallback((err, body, res) => {
+            const callback = client.wrapCallback("POST", (err, body, res) => {
                 err.should.be.deep.equal([{
                     message: "Could not communicate with test-server",
                     code: "COMMUNICATION_ERROR",
+                }]);
+                body.should.be.equal("test");
+                rawRes.should.be.deep.equal(res);
+
+                cb();
+            });
+            callback(new Error(), rawRes);
+        });
+
+        it("should call callback with 'body' and 'res' and application/jose+json Content-Type", (cb) => {
+            const clientPath = path.join(__dirname, "..", "resources", "private-jwkset1");
+            const hwPath = path.join(__dirname, "..", "resources", "public-jwkset1");
+            const clientWithEncryption = new ApiClient("test-username", "test-password", "https://test-server", {
+                clientPrivateKeySetPath: clientPath,
+                hyperwalletKeySetPath: hwPath,
+            });
+
+            const encryption = new Encryption(clientPath, hwPath);
+            const testMessage = {
+                message: "Test message",
+            };
+
+            encryption.encrypt(testMessage).then((encryptedBody) => {
+                const callback = clientWithEncryption.wrapCallback("POST", (err, body, res) => {
+                    expect(err).to.be.undefined();
+                    expect(res).not.to.be.undefined();
+                    body.should.be.deep.equal(testMessage);
+
+                    cb();
+                });
+                const rawRes = {
+                    text: JSON.stringify(encryption.base64Decode(encryptedBody)),
+                    status: 200,
+                    type: "application/jose+json",
+                };
+                callback(undefined, rawRes);
+            });
+        });
+
+        it("should call callback with static error message as 'errors', when Content-Type is wrong", (cb) => {
+            const client = new ApiClient("test-username", "test-password", "test-server");
+
+            const rawRes = {
+                body: "test",
+                status: 200,
+                type: "wrongContentType",
+            };
+
+            const callback = client.wrapCallback("POST", (err, body, res) => {
+                err.should.be.deep.equal([{
+                    message: "Invalid Content-Type specified in Response Header",
                 }]);
                 body.should.be.equal("test");
                 rawRes.should.be.deep.equal(res);