+++ CHANGELOG.md
@@ -3,7 +3,8 @@ Changelog
 
 0.2.0 (in progress)
 -------------------
-
+- Added PayPal account endpoint
+- Added transfer endpoint
 - Added client token endpoint
 - Added Layer 7 encryption
 - Added payment status transition endpoint